Collapse Component_AttriuteModifier to one class
There were two versions - Mult and Flat. Collapsed into one. Next, I
will add conditions (only + to items of a certain label, for example)
which will be the basis of "Pilot has +3 to To Hit for Rifles".

To Hit. Heh. I should do THACO!

No, that's a terrible idea and I should feel bad for thinking it.

+++ MechArena/GameQuery_EntityAttribute.cs
@@ -9,7 +9,7 @@ namespace MechArena
     {
         private bool registeredBase = false;
         private int baseValue = 0;
-        private List<Tuple<int, Entity>> additiveModifiers = new List<Tuple<int, Entity>>();
+        private List<Tuple<double, Entity>> additiveModifiers = new List<Tuple<double, Entity>>();
         private List<Tuple<double, Entity>> multiplicativeModifiers = new List<Tuple<double, Entity>>();
 
         public Entity BaseEntity { get; }
@@ -50,9 +50,9 @@ namespace MechArena
             this.baseValue = value;
         }
 
-        public void AddFlatModifier(int value, Entity source)
+        public void AddFlatModifier(double value, Entity source)
         {
-            this.additiveModifiers.Add(new Tuple<int, Entity>(value, source));
+            this.additiveModifiers.Add(new Tuple<double, Entity>(value, source));
         }
 
         public void AddMultModifier(double value, Entity source)
